test(tiles): split issue 396 regression seams
